SkillSwapProject/requirements.txt
python-dotenv
psycopg2
django 
djangorestframework
django-phonenumber-field
phonenumbers
pillow
pyjwt